pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wadlib"
version = "0.0.42"
description = "Python library to work with id Doom WAD files"
authors = [{ name = "Alex Rembish", email = "alex@rembish.org" }]
license = { text = "BSD-2-Clause" }
requires-python = ">=3.12"
dependencies = ["Pillow>=9.2"]

[project.scripts]
wadcli = "wadlib.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pylint>=3.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["wadlib*"]

# ---------------------------------------------------------------------------
# ruff — linting + formatting (replaces flake8, isort, pyupgrade, etc.)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "UP",       # pyupgrade
    "B",        # flake8-bugbear
    "C4",       # flake8-comprehensions
    "SIM",      # flake8-simplify
    "RUF",      # ruff-specific
]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["wadlib"]

[tool.ruff.format]
quote-style = "double"

# ---------------------------------------------------------------------------
# mypy — strict type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
exclude = ["tests/"]

# ---------------------------------------------------------------------------
# pylint
# ---------------------------------------------------------------------------
[tool.pylint.main]
py-version = "3.12"

[tool.pylint.design]
max-attributes = 16

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "R0903",  # too-few-public-methods
    "W0511",  # fixme — TODO comments are intentional
    "C0414",  # useless-import-alias — used for explicit re-exports (mypy)
    "R0801",  # duplicate-code — CLI commands share similar boilerplate by design
]

[tool.pylint.format]
max-line-length = 100

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q --cov=wadlib --cov-report=term-missing -m 'not slow'"
markers = [
    "slow: marks tests as slow (deselected by default; run with -m slow or -m '')",
]

[tool.coverage.run]
branch = true
source = ["wadlib"]

[tool.coverage.report]
exclude_lines = ["if TYPE_CHECKING:"]
